requirements.txt
gtts>=2.1.1
praw>=7.0.0
python-ffmpeg>=1.0.11
